requirements.txt
Django==1.11.5
django-phonenumber-field==1.1.0
py-moneyed==0.7.0
django-money==0.11.4